+++ database/migrations/2020_04_05_062520_create_replies_table.php
@@ -17,8 +17,8 @@ class CreateRepliesTable extends Migration
             $table->bigIncrements('id');
             $table->text('body');
 
-            $table->integer('question_id')->unsigned();
-            $table->integer('user_id')->unsigned();
+            $table->bigInteger('question_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
 
             $table->foreign('question_id')->references('id')->on('questions')->onDelete('cascade');
             $table->timestamps();
